Get built-in tests passing (add escape-by-default to jellies)
Running `mvn test` with the more recent Jenkins dependencies seems
to cause some built-in tests to get run, which is nice.  However,
one of the tests appeared to be validating the jelly files in the
plugin in some way, and they were failing validation because it
was validating that they should contain a line like this:

    <?jelly escape-by-default='true'?>

This commit simply adds that line to all of the jelly files.  Tests
pass as a result.