Add skeleton test for new Gatling "step"
This commit just adds a skeleton test that we will be able to use
to test a new Gatling pipeline "step" when we add it.  The commit
also adds dependencies on the Jenkins plugins that are required
to get the test running.

The Gatling step itself will be introduced in a subsequent commit.