Add a skeleton for a 'step', and update test
This commit adds a pipeline "step", which allows you to call, e.g.,
`gatlingArchive()` in your pipeline script.  The implementation
of the step doesn't do anything yet, but the test is updated to
validate that the syntax is indeed now allowed in a pipeline script.

After this commit it is also possible to go into the Jenkins GUI,
go to the configuration screen for a pipeline job, open up the
snippet generator, and select "gatlingArchive" from the dropdown.