Port Publisher over to use SimpleBuildStep API
This commit ports the GatlingPublisher class to use the newer
`SimpleBuildStep` API, instead of the old `AbstractBuild` API.

After this commit, it is possible to run the Gatling publisher
in a pipeline script, by using the following syntax:

    step([$class: 'GatlingPublisher', enabled: true])

This step will also show up in the groovy snippet generator now, if you
select `step` from the first dropdown.

After this commit the plugin will successfully archive the reports,
and will show the "Gatling" link in the sidebar on the individual
build page.  It won't show the graph of historical data on the
main project/job page, though.  Still need to sort that out.