Modify GatlingArchiverStep to call GatlingPublisher
This commit modifies the implementation of the GatlingArchiverStep
so that it is no longer a noop, but actually calls the GatlingPublisher.

This means that you can include a line like this in your Jenkins Pipeline
script:

    gatlingArchive()

And the effect will be the same as if you'd added the Gatling post-build
action to your freestyle job; that is, you'll see the Gatling link appear
in the sidebar of both the Project and Build pages, the reports will be
available, the graphs will be visible, etc.